+++ Changelog.md
@@ -1,6 +1,7 @@
 # Unreleased
 
 - Add implementations for `fmt::Pointer`, `PartialEq`, `Eq`, `PartialOrd`, `Ord` and `Hash`.
+- Update `very_unstable` feature to latest nightly
 
 # 0.5.1 – 2023-06-24